requirements.txt
wheel
torch
transformers
openai
python-dotenv
docopt